Currently 3.5.1 will not work
When using DNS names within the config file (what we're relying on because of Kubernetes services), we need to modify the file for each Zookeeper instance, replacing it's DNS name with 0.0.0.0 in the server connection list, for example:

`server.1=zookeeper-1:2888:3888`
will become
`server.1=0.0.0.0:2888:3888`
but only on server.1, and likewise for each server on their respective configs.

Where we're running into trouble is that in 3.5.X, Zookeeper introduces dynamic configs (initially yay!), which sync configs amongst members of the cluster.  That is trouble because we end up with the leader propigating it's 0.0.0.0 route to the rest of the members and then they can't connect to it.

For now we solve this by rolling back to 3.4.8 and waiting for them to either:
- expose some way to choose which address we can bind to for leader elections so we don't receive the error
- outsmart  the syncing of the dynamic configs (this would be best because we do like the ability to add members after the fact)